+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- Adds `interceptCallback` parameter to allow for custom interception of the callback URL
+
 ## 0.0.3
 
 - Supports discovery